requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.4
async-timeout==3.0.1
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.0
channels==3.0.3
channels-redis==3.2.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
daphne==3.0.2
Django==3.2.4
django-allauth==0.44.0
hiredis==2.0.0
hyperlink==21.0.0
incremental==21.3.0
kombu==5.1.0
msgpack==1.0.2
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests-oauthlib==1.3.0
service-identity==21.1.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
vine==5.0.0
zope.interface==5.4.0
